added async main/renderer communication handling
